Themes almost work (with a few bugs); New system themes; Cleaned up code;

+++ apps/web/components/admin/page-editor/edit-widget.tsx
@@ -36,6 +36,10 @@ export default function EditWidget({
     const actualWidget = widgets[widget.name];
 
     const onDeleteWidget = () => {
+        if (!widget.deleteable) {
+            return;
+        }
+
         if (deleteConfirmation) {
             onDelete(widget.widgetId);
         } else {
